fastapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.3.2
